Refactor: Course Controller

+++ controllers/course.controller.js
@@ -4,6 +4,7 @@ async function createCourse(req, res) {
 
     try {
         const { title, description } = req.body
+
         const createdCourse = await Course.create({
             title,
             description,
@@ -11,18 +12,22 @@ async function createCourse(req, res) {
         })
 
         res.status(201).json(createdCourse)
-    } catch (error) {
-        res.status(500).json({ message: error.message })
+    }
+    catch (error) {
+        if (err.name === 'ValidationError') {
+            return res.status(400).json({ message: err.message })
+        }
 
+        res.status(500).json({ message: error.message })
     }
 }
 
 async function allCourses(req, res) {
     try {
-        const getAllCourses = await Course.find({owner: req.user._id})
+        const getAllCourses = await Course.find({ owner: req.user._id })
         res.status(200).json(getAllCourses)
-
-    } catch (error) {
+    } 
+    catch (error) {
         res.status(500).json({ message: error.message })
     }
 }
@@ -30,7 +35,7 @@ async function allCourses(req, res) {
 async function getCourse(req, res) {
     try {
         const getOneCourse = await Course.findById(req.params.courseId)
-        
+
         if (!getOneCourse) {
             return res.status(404).json({ message: 'no course found, add yours!' })
         }
@@ -44,9 +49,21 @@ async function getCourse(req, res) {
 
 async function updateCourse(req, res) {
     try {
+
+        const foundCourse = await Course.findById(req.params.courseId)
+
+        if (foundCourse.owner != req.user._id) {
+            return res.status(403).json({ message: 'You are not authorized to update this course details' })
+        }
+
         const { title, description } = req.body
-        const updatedCourse = await Course.findByIdAndUpdate(req.params.courseId, { title, description }, { new: true })
-        res.status(200).json(updatedCourse)
+        foundCourse.title = title
+        foundCourse.description = description
+        await foundCourse.save()
+
+        // const updatedCourse = await Course.findByIdAndUpdate(req.params.courseId, { title, description }, { new: true })
+        // res.status(200).json(updatedCourse)
+        res.status(200).json(foundCourse)
 
     } catch (error) {
         res.status(500).json({ message: error.message })
@@ -55,16 +72,23 @@ async function updateCourse(req, res) {
 
 async function deleteCourse(req, res) {
     try {
+
+        const foundCourse = await Course.findById(req.params.courseId)
+
+        if (foundCourse.owner != req.user._id) {
+            return res.status(403).json({ message: 'You are not authorized to update this course details' })
+        }
+
         const deletedCourse = await Course.findByIdAndDelete(req.params.courseId)
-        
+
         if (!deletedCourse) {
             return res.status(404).json({ message: 'no course found, add yours!' })
         }
-        
+
         res.status(204).json(deletedCourse)
+
     } catch (error) {
         res.status(500).json({ message: error.message })
-
     }
 }
